[Feature] UI, Presenter 구성
