+++ Team_Immortal Sprouts_Pummel Party/Assets/Art/Enviroments/BoardGame/Camera/New Render Texture.renderTexture.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 27b858729c3748f4aa52c17f80f4d0fb
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 8400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
