+++ Team_Immortal Sprouts_Pummel Party/Assets/Art/Sprites/MiniGameInfo.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 97d69c0cd3a64d7418c4c9e6082e0aab
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
