+++ Team_Immortal Sprouts_Pummel Party/Assets/Scripts/2_Boardgame Scene/Turn/BoardGameFrameWork.cs
@@ -1,0 +1,17 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.Events;
+
+public class BoardGameFrameWork : MonoBehaviour
+{
+    public UnityEvent OnFirtstStartBoardGame;
+
+    public UnityEvent OnStartBoardGame;
+
+    public UnityEvent TurnOnPlayer;
+
+    public UnityEvent TurnOffPlayer;
+
+    public UnityEvent OnAllPlayerTurnOff;
+}
