+++ Team_Immortal Sprouts_Pummel Party/Assets/Scripts/2_Boardgame Scene/Turn/BoardGameFrameWork.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 86fed0953078f7f43ae7bd890fc9061c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
