+++ Team_Immortal Sprouts_Pummel Party/Assets/Scripts/2_Boardgame Scene/Turn/BoardGameManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0e5d13b4ee083df4a8b893fa500c73cf
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
