+++ Team_Immortal Sprouts_Pummel Party/Assets/Scripts/2_Boardgame Scene/UI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f1a9b1229fcd27241bef6fb1ae02378d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
