+++ Team_Immortal Sprouts_Pummel Party/Assets/Scripts/2_Boardgame Scene/UI/BoardGamePresenter.cs
@@ -1,0 +1,9 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.Events;
+
+public class BoardGamePresenter : MonoBehaviour
+{
+    public UnityEvent OnMiniGameStart;
+}
