+++ Team_Immortal Sprouts_Pummel Party/Assets/Scripts/2_Boardgame Scene/UI/BoardGamePresenter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6fb48dfdfc86145428fb66ca2a7ec479
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
